requirements.txt
numpy
pandas
azure-storage-blob